requirements.txt
Flask
Flask-SQLAlchemy
MySQL-python
gunicorn
eru-py
redis
